State management / file org restructuring per code review
