+++ apps/react/demo/src/data/examples.ts
@@ -2,10 +2,10 @@ import {
   ButtonExample,
   TypographicSpecimen,
 } from "../ui/Example/common/index.js";
-import type { ShowcaseExampleOpts } from "../ui/Example/index.js";
+import type { ShowcaseExample } from "../ui/Example/index.js";
 import DEFAULT_CONTROLS from "./defaultSettings.js";
 
-export const SHOWCASE_EXAMPLES: ShowcaseExampleOpts[] = [
+export const SHOWCASE_EXAMPLES: ShowcaseExample[] = [
   {
     name: "Typographic Specimen",
     description: "A typographic specimen with configurable font settings",
