+++ apps/react/demo/src/ui/Example/hooks/useConfig.ts
@@ -1,11 +1,11 @@
 import { useContext } from "react";
-import Context from "../../Context.js";
-import type { ProviderValue } from "../../Provider/index.js";
+import Context from "../Context.js";
+import type { ContextOptions } from "../types.js";
 
 /**
  * Hook to access the config object from the nearest ConfigProvider.
  */
-const useConfig = (): ProviderValue => {
+const useConfig = (): ContextOptions => {
   const context = useContext(Context);
   if (!context) {
     throw new Error("useConfig.ts must be used within a ConfigProvider");
