--- apps/react/demo/src/ui/Example/hooks/useConfig/index.ts
@@ -1,1 +1,0 @@
-export { default as useConfig } from "./useConfig.js";
