+++ apps/react/demo/src/ui/Example/hooks/useExampleControls/index.ts
@@ -1,1 +1,2 @@
 export { default as useExampleControls } from "./useExampleControls.js";
+export * from "./types.js";
